Change the field for the size to Interger Field to fix size sorting

+++ products/models.py
@@ -24,7 +24,7 @@ class Product(models.Model):
     brand = models.CharField(max_length=254, null=True, blank=True)
     description = models.TextField()
     price = models.DecimalField(max_digits=6, decimal_places=2)
-    size = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)  # Use DecimalField for size
+    size = models.IntegerField(null=True, blank=True) 
     image = models.ImageField(upload_to='images/', null=True, blank=True) 
 
     def __str__(self):
